+++ Runtime/Services/LocalizationService/DummyLocalizationChangeNotifier.cs
@@ -14,6 +14,7 @@ namespace LocalizationService
 
         public void Initialize()
         {
+            Changed?.Invoke();
             //ignore
         }
     }