+++ Runtime/Services/LocalizationService/DummyLocalizationService.cs
@@ -1,11 +1,18 @@
 using System;
 using LocalizationService.RuntimeLocalization;
-using TMPro;
+using UnityEngine;
 
 namespace LocalizationService
 {
     public class DummyLocalizationService : ILocalizationService
     {
+        private readonly DummyLocalizationChangeNotifier m_localizationChangeNotifier;
+
+        public DummyLocalizationService()
+        {
+            m_localizationChangeNotifier = new DummyLocalizationChangeNotifier();
+        }
+
         public string GetString(string key, string tableName = "Default Localization Table", params object[] arguments)
         {
             return key.ToUpper();
@@ -16,13 +23,15 @@ namespace LocalizationService
             return key.ToUpper();
         }
 
-        public IRuntimeLocalizedText ConvertToUnityRuntimeLocalizedTMP(TMP_Text text, string key, string table = "Base String Table",
+        public IRuntimeLocalizedText ConvertToRuntimeLocalizedText(PerformTextSet performTextSet, string key, GameObject linkedObject,
+            string table = "",
             Func<object[]> argumentsFunc = null,
             string format = "{0}")
         {
-            RuntimeLocalizedTMP instance = text.gameObject.AddComponent<RuntimeLocalizedTMP>();
-            instance.Setup(text, key, table, format, argumentsFunc);
-            instance.Initialize(this, new DummyLocalizationChangeNotifier());
+            RuntimeLocalizedTextSetAction instance = new();
+            instance.Setup(performTextSet, key, table, format, argumentsFunc);
+            instance.Initialize(this, m_localizationChangeNotifier);
+            instance.LinkTo(linkedObject);
             return instance;
         }
     }