+++ Runtime/Services/LocalizationService/LocalizationConfiguration.cs
@@ -1,0 +1,8 @@
+namespace LocalizationService
+{
+    public class LocalizationConfiguration
+    {
+        public string DefaultTableName { get; set; } = "Default Localization Table";
+        public static LocalizationConfiguration Default => new LocalizationConfiguration();
+    }
+}