+++ Runtime/Services/LocalizationService/Unity/UnityLocalizationService.cs
@@ -16,22 +16,25 @@ namespace LocalizationService.Unity
         private const char SplitChar = ';';
         private const string ErrorPlaceholder = "No translation";
         private readonly ILocalizationChangeNotifier m_notifier;
+        private readonly LocalizationConfiguration m_localizationConfiguration;
 
-        public UnityLocalizationService(ILocalizationChangeNotifier localizationChangeNotifier)
+        public UnityLocalizationService(ILocalizationChangeNotifier localizationChangeNotifier,
+            LocalizationConfiguration localizationConfiguration)
         {
+            m_localizationConfiguration = localizationConfiguration;
             m_notifier = localizationChangeNotifier;
             m_notifier.Initialize();
         }
 
-        public string GetString(string key, string tableName = "Default Localization Table", params object[] arguments)
+        public string GetString(string key, string tableName = "", params object[] arguments)
         {
             return Internal_GetString(key, tableName, arguments);
         }
 
-        public string GetStringByIndex(string key, int index, string tableName = "Default Localization Table",
+        public string GetStringByIndex(string key, int index, string table = "",
             params object[] arguments)
         {
-            string[] split = Internal_GetString(key, tableName, arguments).Split(SplitChar);
+            string[] split = Internal_GetString(key, table, arguments).Split(SplitChar);
 
             if (split.Length != 0 && split.Length > index)
                 return split[index];
@@ -40,24 +43,28 @@ namespace LocalizationService.Unity
             return ErrorPlaceholder;
         }
 
-        public IRuntimeLocalizedText ConvertToUnityRuntimeLocalizedTMP(TMP_Text text, string key,
-            string table = "Base String Table", Func<object[]> argumentsFunc = null, string format = "{0}")
+        private string GetFixedTableName(string input)
         {
-            RuntimeLocalizedTMP component = text.gameObject.AddComponent<RuntimeLocalizedTMP>();
-            component.Setup(text, key, table, format, argumentsFunc);
-            component.Initialize(this, m_notifier);
-            return component;
+            if (string.IsNullOrEmpty(input)) input = m_localizationConfiguration.DefaultTableName;
+            return input;
         }
 
-
-        private static string Internal_GetString(string key, string table, params object[] arguments)
+        public IRuntimeLocalizedText ConvertToRuntimeLocalizedText(PerformTextSet performTextSet, string key, GameObject linkedObject,
+            string table = "",
+            Func<object[]> argumentsFunc = null,
+            string format = "{0}")
         {
-            if (string.IsNullOrEmpty(key))
-            {
-                Debug.Log($"Given empty key {key}");
-                return ErrorPlaceholder;
-            }
+            table = GetFixedTableName(table);
+            RuntimeLocalizedTextSetAction instance = new();
+            instance.Setup(performTextSet, key, table, format, argumentsFunc);
+            instance.Initialize(this, m_notifier);
+            instance.LinkTo(linkedObject);
+            return instance;
+        }
 
+        private string Internal_GetString(string key, string table, params object[] arguments)
+        {
+            table = GetFixedTableName(table);
             string found = LocalizationSettings.StringDatabase.GetLocalizedString(table, key, arguments);
             return found;
         }